+++ src/main/java/Domain/Store/Inventory/ProductDTO.java
@@ -9,13 +9,13 @@ public class ProductDTO {
     private Integer productID;  // The unique ID of the product
     private String name;        // The name of the product
     //private String desc;        // The description of the product
-    private int price;          // The price of the product
+    private double price;          // The price of the product
    // private int quantity;       // The quantity of the product available in the inventory
 
     public ArrayList<String> categories; // The category that a product is related to
 
 
-    public ProductDTO(String storeID, Integer productID, String name, int price) {
+    public ProductDTO(String storeID, Integer productID, String name, double price) {
         this.storeID = storeID;
         this.productID = productID;
         this.name = name;
